+++ packages/atom.io/src/internal/is-default.ts
@@ -1,5 +1,3 @@
-import HAMT from "hamt_plus"
-
 import type { Store } from "."
 import { target, IMPLICIT, traceAllSelectorAtoms } from "."
